Move multisig helper to client helpers
